Silence CS0675 warnings, add README badges, fix wrong license claim
CS0675 (bitwise-or on a sign-extended operand) fired 24 times in CI for
SetBit1/SetBit1Copy on sbyte/short, where "value" was the direct left
operand of "|" and got sign-extended through the implicit int promotion.
The result was already correct - the final narrowing cast back to
sbyte/short discards exactly the bits sign extension added - but the
original code sidestepped the warning entirely by casting to the unsigned
counterpart first (e.g. (UInt16)value | ...), and I'd dropped that when
simplifying these methods. Restored it: 0 warnings, same behavior,
confirmed by the existing SetBit1/SetBit1Copy tests plus a full local
build (was 1041 warnings incl. 24x CS0675, now 1029 with byte-for-byte
identical codegen for these methods since the intermediate widening was
always optimized away).

README: added the standard badge row (NuGet version, NuGet downloads,
GitHub Actions build status, license), matching the convention already
used in Tedd.AsyncLock. Also fixed the License section, which I'd written
as "MIT" without checking - this repo's LICENSE file is actually LGPL-3.0.

No functional change and no version bump, so this does not trigger a new
NuGet publish (the workflow only publishes on a version bump).

+++ src/Tedd.BitUtils/BitUtilsCopyExtensions.cs
@@ -78,13 +78,13 @@ public static partial class BitUtilsCopyExtensions
     /// <param name="pos">Zero based bit position counted from the least significant bit.</param>
     /// <returns>Modified copy.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static sbyte SetBit1Copy(ref this sbyte value, int pos) => (sbyte)(value | (1 << pos));
+    public static sbyte SetBit1Copy(ref this sbyte value, int pos) => (sbyte)((byte)value | (1 << pos));
     /// <inheritdoc cref="SetBit1Copy(ref sbyte, int)"/>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static byte SetBit1Copy(ref this byte value, int pos) => (byte)(value | (1 << pos));
     /// <inheritdoc cref="SetBit1Copy(ref sbyte, int)"/>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static short SetBit1Copy(ref this short value, int pos) => (short)(value | (1 << pos));
+    public static short SetBit1Copy(ref this short value, int pos) => (short)((ushort)value | (1 << pos));
     /// <inheritdoc cref="SetBit1Copy(ref sbyte, int)"/>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static ushort SetBit1Copy(ref this ushort value, int pos) => (ushort)(value | (1 << pos));
